agents/okr_parser_agent/requirements.txt
fastapi
uvicorn
langchain
openai
faiss-cpu
python-dotenv 